using original get_action instead of implementing get_generic_action
Using the original get_action method simplifies the code. Though with a very rare missconfiguration we risk a strange error: If the generic_action configured with the ModelAdmin is not a callable and not found as ModelAdmin method but has by chance the same name as an action defined with the AdminSite, this action would be passed as generic-action. But since this is a rare and unlikely situation produced by a misconfiguration in the first place, we take the risk with the benefits.